requirements.txt
python-telegram-bot
python-twitch-client